Filter out applications in the runtime based on application type
