+++ chart/compass/charts/director/config.yaml
