+++ chart/compass/charts/director/values.yaml
@@ -23,3 +23,5 @@ staticGroupNames:
   - "mps-superadmin"
   - "runtimeAdmin"
   - "runtimeNamespaceAdmin"
+
+applicationHideSelectors:
