+++ components/director/pkg/config/errors.go
@@ -1,0 +1,5 @@
+package config
+
+import "github.com/pkg/errors"
+
+var PathNotFoundError = errors.New("specified path not found in configuration")
