+++ components/director/pkg/config/testdata/invalid.yaml
