+++ components/director/pkg/config/testdata/valid.yaml
@@ -1,0 +1,29 @@
+graphql:
+  query:
+    runtime: "runtime:get"
+    application: "application:get"
+
+  mutation:
+    createApplication: ["application:create","global:create"]
+    deleteApplication: "application:delete"
+    updateApplication: [1,2,3]
+    empty:
+
+clientCredentialsRegistrationScopes:
+  runtime:
+    - "runtime:read"
+    - "runtime:write"
+    - "application:read"
+  application:
+    - "application:read"
+    - "application:write"
+  integration_system:
+    - "application:read"
+    - "application:write"
+    - "runtime:read"
+    - "runtime:write"
+
+applicationHideSelectors:
+  applicationType:
+    - "Test/App"
+    - "Work In Progress"
