--- components/director/pkg/scope/testdata/invalid2.yaml
@@ -1,8 +1,0 @@
-queries:
-  runtime: "runtime:get"
-  application: "application:get"
-  applications: "application:get"
-mutations:
-  createApplication: ["application:create","global:create"]
-  deleteApplication: "application:delete"
-  updateApplication: [3,2,1]
