--- components/director/pkg/scope/testdata/invalid3.yaml
@@ -1,7 +1,0 @@
-queries:
-  runtime: "runtime:get"
-  application: 3
-  applications: "application:get"
-mutations:
-  createApplication: ["application:create","global:create"]
-  deleteApplication: "application:delete"
