--- components/director/pkg/scope/testdata/valid.yaml
@@ -1,8 +1,0 @@
-queries:
-  runtime: "runtime:get"
-  application: "application:get"
-mutations:
-  createApplication: ["application:create","global:create"]
-  deleteApplication: "application:delete"
-  updateApplication: [1,2,3]
-  empty:
